+++ src/WelcomeController.php
@@ -1,26 +1,28 @@
 <?php
+
 /**
- * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @copyright Copyright (c) 2014 Zend Technologies USA Inc. (http://www.zend.com)
+ * @see       https://github.com/laminas-api-tools/api-tools-welcome for the canonical source repository
+ * @copyright https://github.com/laminas-api-tools/api-tools-welcome/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas-api-tools/api-tools-welcome/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZF\Apigility\Welcome;
+namespace Laminas\ApiTools\Welcome;
 
-use Zend\Mvc\Controller\AbstractActionController;
-use Zend\View\Model\ViewModel;
+use Laminas\Mvc\Controller\AbstractActionController;
+use Laminas\View\Model\ViewModel;
 
 class WelcomeController extends AbstractActionController
 {
     public function welcomeAction()
     {
         $viewModel = new ViewModel();
-        $viewModel->setTemplate('zf-apigility-welcome/welcome');
+        $viewModel->setTemplate('api-tools-welcome/welcome');
 
         return $viewModel;
     }
 
     public function redirectAction()
     {
-        return $this->redirect()->toRoute('zf-apigility/welcome');
+        return $this->redirect()->toRoute('api-tools/welcome');
     }
 }