+++ CHANGELOG.md
@@ -28,7 +28,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) adds support for PHP 7.0 - 7.4.
+- [zfcampus/zf-apigility-welcome#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) adds support for PHP 7.0 - 7.4.
 
 ### Changed
 
@@ -40,9 +40,9 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Removed
 
-- [#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) removes support for PHP versions prior to 5.6.
+- [zfcampus/zf-apigility-welcome#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) removes support for PHP versions prior to 5.6.
 
-- [#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) removes support for HHVM.
+- [zfcampus/zf-apigility-welcome#6](https://github.com/zfcampus/zf-apigility-welcome/pull/6) removes support for HHVM.
 
 ### Fixed